Initial work for extending loading
