+++ mantidimaging/core/utility/size_calculator.py
@@ -74,5 +74,5 @@ def full_size_MB(shape=None, axis=None, dtype=None):
     return to_MB(full_size(shape, axis), dtype)
 
 
-def number_of_images_from_indices(start, end, step):
-    return int((end - start) / step) if step != 0 else 0
+def number_of_images_from_indices(start, end, increment):
+    return int((end - start) / increment) if increment != 0 else 0
