+++ mantidimaging/gui/windows/main/view.py
@@ -8,7 +8,7 @@ from mantidimaging.core.data import Images
 from mantidimaging.core.utility.version_check import find_if_latest_version
 from mantidimaging.gui.mvp_base import BaseMainWindowView
 from mantidimaging.gui.windows.filters import FiltersWindowView
-from mantidimaging.gui.windows.main.load_dialog import MWLoadDialog
+from mantidimaging.gui.windows.main.load_dialog.view import MWLoadDialog
 from mantidimaging.gui.windows.main.presenter import MainWindowPresenter
 from mantidimaging.gui.windows.main.presenter import Notification as PresNotification
 from mantidimaging.gui.windows.main.save_dialog import MWSaveDialog
